requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-dajax==0.9.2
django-dajaxice==0.7
django-on-heroku==1.1.2
gunicorn==20.0.4
idna==2.10
paho-mqtt==1.5.1
psycopg2-binary==2.8.6
pytz==2020.4
requests==2.25.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
